Style: 함수 분리
