+++ src/components/RhythmList.tsx
@@ -84,9 +84,7 @@ export default function RhythmList({ selectedDate }: RhythmListProps) {
             <StyledRhythmListYear>{year}</StyledRhythmListYear>
           </StyledRhythmListYearMonthWrapper>
         </StyledRhythmListDateWrapper>
-        <StyledWeatherBox>
-          <Weather />
-        </StyledWeatherBox>
+        <Weather />
       </StyledRhythmListHead>
       <StyledRhythmTableBox>
         {isLoading ? (
@@ -262,7 +260,4 @@ const StyledRhythmListCircleButton = styled.button`
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-`;
-const StyledWeatherBox = styled.div`
-  height: 3.5rem;
 `;
